snippets.bash: remove \n when copying to clipboard
read.js is invalid JS because it contains a string with raw newlines. This is
not a problem for the bookmarklet, because the newlines get stripped by the
generate-bookmarks-html script. I leave it as-is in the source code to allow
easier editing. I could use E4X for a multiline string, but then I would lose
support for other browsers.