Readable++: add default colours
I long debated doing this, because I did not want to override the user's own
preferences. Today, I finally caved and gave in to my opponent, which was I
myself. Whodathunkit?!

Anyway, the reason for the change is that I added background colours for
things like CODE/PRE and data tables, so not having default colours meant things
could get ugly: for instance, a yellow CODE background (from this bookmarklet)
on a black BODY background with white text (from the user's preferences) would
lead to unreadable text.