generate-bookmarks-html: strip trailing backslash
This makes the "Readable++" bookmarklet work again. I had not noticed because
I was not using the generated version.

I know this is a naive approach, but for now I will ass-u-me there are no
trailing backslashes that are escaped by a backslash themselves. ("Yo, dawg,
I heard you liked backslashes so I put a backslash before your backslash so
you can escape what is used to escape.")