Readable++: show all images etc. when hovering
This makes for a more pleasant reading experience. You can still choose to
have the images be less obtrusive by not hovering over BODY. Note that on
larger screens, you can still use the scroll wheel when over HTML, because
BODY is only 52 ems wide.

I will take this for a spin on my iPhone to see if tapping the BODY will
trigger the :hover and make the images etc. fully opaque.