Google Maps: force "classic" version
I do not like the new version, and I have not yet found a way to specify
the transportation mode ("dirflg" in the "classic" version). The URLs are
much less human-readable now. I need to look into the encoding of the "data"
parameter, but until then, make the "classic" version work again. Yay!