Readable++: scroll to title header
If there is an element that looks like a heading and whose text appears in the
page title, ass-u-me it is the start of the main content of the page.

Seems simple enough, and with a few checks in place it appears to be quite
useful so far. And if no suitable header was found, it falls back to the
original content element detection.