Readable++: choose most suitable header if several
"Most suitable header" being: "the header whose text appears nearest the
beginning of the page title. This is what most sites seem to do, in line with
SEO recommendations.

Sites that use "Site Title | Page Title" *and* have multiple matching headers
(i.e. at least one with the page title and one with the site title) will be at
a disadvantage, because the header with the site title will be considered the
main header element.