*: fall back when data: URIs cannot be opened
Firefox 57 blocks navigation to data: URIs in the top-level window, so we can
no longer set window.location to "data:text/html,foo". To work around this,
wait for a bit (long enough for the navigation to succeed in browsers that do
allow opening top-level data: URIs) and rewrite the current document. I tried
modifying document.documentElement.outerHTML, but that was not allowed
("NoModificationAllowedError"). I then tried setting innerHTML, but that was
buggy when opening a bookmarklet in a new tab using Alt+Enter. The content was
there in the DOM, but it never showed up on the screen.

Good old document.open() clears the document, so a subsequent document.write()
effectively replaces all content. Old school.

See https://www.fxsitecompat.com/en-CA/docs/2017/data-url-navigations-on-top-level-window-will-be-blocked/