read: rm known-bad elements
I probably should have added this as a site-specific bookmarklet for standaard.be,
but that would cost a lot of extra clicks/taps on Firefox mobile. One of the main
uses of Readable++ is on standaard.be, so this saves me a lot of time. After all,
they still are *my* bookmarks. Mine. MIIIIIIIIIIIIIIIIIIIIIIIIINE!