shide: avoid multiline info string (for generator)
generate-bookmarks-html replaces all newlines by regular spaces. Rather
than fixing that old script to be aware of multiline `` strings in
JavaScript, simply work around the problem by using an array with one
element for each line. Not as clean, but less work (and no
fall-out/regression due to the impact on other bookmarklets).