nocookie: recurse at every query
Because the list of documents can change e.g. when clicking on a “Customize” button
(because a new IFRAME was opened, for instance), going through all documents once can
miss out on dynamically loaded elements.

This new way of always getting all documents is not exactly efficient, but I could
add caching if it turns out to be problematic for speed.