*: remove `console` restoration
Trying to make the browser’s original `console` available to the
bookmarklet led to messages not being logged at all, especially when
unloading/navigating to another page.

Let’s try it like this; if I get too many errors about `console.log`
etc. not being a function (because it was deleted/sabotaged by the
page), I can always revert. Or change it to:

    `typeof console?.log === 'function' && console.log('…')`

and simply hope that points to the original.