prev/next: look for dates in URL without %20 etc.
Using “next” on a path with `%201979` would go to `%2019710` because the
code thought it was `2019-7-9`.

Better still would be to only allow zero-less single digits for month
and day iff there were separators, because it is not reasonable to
ass-u-me that `201979` means `2019-07-09` is not a reasonable way to
write `July 9, 2019`.