read: use multi-line string for style sheet
Long overdue for a more thorough modernisation, methinks.

This is the oldest bookmarklet in this repository, and it shows. The
first commit was 13 years ago, but I had been using this bookmarklet
before the repository existed. I think I created the original version in
2008 or some time around then.

COOL STORY, BRO!

(You know, 2008: when “COOL STORY, BRO!” was not yet a *completely*
outdated way to react to a boring ancedote.)