2*: stop using string replacement to create clones
A follow-up to commit 7afe5db3ceb9.

Using string replacement to replace `en`/`English`/`this page in
English`/… in the original `2en` bookmarklet to create the versions for
other languages like `2nl`/`2fr`/… was error-prone and inflexible.

By specifying the language-specific bits as function parameters, it is
easier to add new functionality like supporting multiple language codes
for the same language family (e.g. Norsk/Bokmål/Nynorsk) and a non-fixed
number of `this page in <Language>` variants.