generate-bookmarks-html: ensure unique URLs
When importing into Firefox, the `eimg` bookmarklet did not get the
keyword. Turns out the keywords (`A@SHORTCUTURL`) are linked to the
URLs (`A@HREF`), and the last keyword linked to a URL is used for all
other times the URL is seen. So `search/eimg.js` would get the keyword
`img` because that was seen last (using the transclusion from
`search/img.js`), even though the bookmarks HTML says `eimg`.

By appending the original filename after `javascript:(…)`, the
bookmarklet URL becomes unique.